Fixed back button issue near SearchView!
While clicking the back button near SearchView initially the query in the SearchView is not cleared so that the recyclerView of the MainActivity is still showing the results of the search in the SearchView. So, I just fixed the issue with it so after clicking the back button, the query in the SearchView is cleared and the MainActivity shows the default list of users without any search.